fix edit and delete functionalities
